Add identifier to export representations used for icons [core]
This adds an explicit icon reference to the representation links to
avoid the frontend having to fiddle with the keys to provide the icons.

Instead, the link object may receive an optional identifier to uniquely
identify the link. I deem that useful in other situations, so I added it
directly into HalLink.