[25073] Force table refresh when updating hierarchies
Currently, altering the hierarchy of a work package does not properly
update the table. For example, removing a parent of w1 that isnt visible in
the table causes w1 to refresh in the table (due to cache service), but
the entire table isn't refreshed.

We cannot refresh the entire table for cache service changes, since that
happens quite often in the timeline mode and would cause a peformance
impact.

On the other hand, we don't have the information on _what_ changed in w1
when we subscribe to the cache service.

I reckon that two solutions exist:

1. Whenever the hierachy changes through user interaction, forcefully
refresh the entire table. This is consistent to the current way of
reloading the table in many situations and will get us the newest data.

2. Provide a frontend representation of the hierarchy of work packages
and update that according to the user's inputs. This puts a lot more
effort into the frontend at the benefit of immediate updates.

The issue I currently see with solution (2) are that we use the
`ancestors` information of a work package to display the hierarchy,
while we modify `parent` and `children` in each work package.
We would need a tree representation of `ancestors` as well as parent and
children maps for fast lookup.

Thus, this PR takes the quick route to fix the issue and forces the page
reload. Since moving from show->list will no longer spawn a refresh
after #5397 is done, we can simply clear the checksum by listening to
the same events.