Fix switching to types with new _links added
When switching to a type that has a multi-select CF, its values are sent
through `payload._links` of the form, however only embedded properites
of the form are taken over.

We can re-use `$initialize` of the HalResource if we merge both sources,
which is what this PR proposes. It also adds a spec to test moving to a
multi-select CF.