[25245] Show loading indicator for single WP save in timeline
After moving a single element in the timeline, the table has to be
reloaded for potential subsequent changes (e.g., movement in relations,
parents, children, etc).

By showing a loading indicator for the single WP save action, we avoid
having a gap between the action and the loading indicator for the table.

On the other hand, this is not shown for inline-editing (unchanged, the
work packages are loaded in the background after saving the line).

Quick fix for:
https://community.openproject.com/projects/openproject/work_packages/25245