truncate wp subject on split and full view (#5498)
I don't consider this to be a permanent fix at least for the full view. The subject currently wraps together with the back button which I don't think to be desirable. It would IMO be better to leave them all in one row and let the subject wrap to the bottom as it becomes longer.

[ci skip]