[25266] Let the timeline container keep the overflow width
Since all global elements are positioned absolute, we don't know the actual width from within the timeline container.

This causes all elements to be of 100% width, and beyond that not responding to events.

However, the browser gives us the total size of all overflowing elements in the container through `scrollWidth`, which we can simply set when drawing.

To use that, we only have to reset its width before rendering, only to get the new/correct scrollWidth afterwards.