[25345] Remove [data-ui-route], instead use ui-router ui-sref
The query menu items are rendered by rails but should be processed by
angular, so the solution was to find clicks on special links marked with
a [data-ui-route] attribute and pass the link to $location.

1. This fails on installations with relative URL, since a base tag is
set and everything passed to $location will be ASSUMED relative to that
base. Passing `/myrelative/work_packages` will result in
`/myrelative/myrelative/work_packages` if the base is `/myrelative`.

2. ui-router already provides a way to build links from state names with
parameters: `ui-sref`. We can simply use that from rails to render the
link.

https://community.openproject.com/projects/openproject/work_packages/25345