[25345] Pass ui-router state in [data-ui-route]
The query menu items are rendered by rails but should be processed by angular, so the solution was to find clicks on special links marked with
a [data-ui-route] attribute and pass the link to $location.

1. This fails on installations with relative URL, since a base tag is set and everything passed to $location will be ASSUMED relative to that
base. Passing `/myrelative/work_packages` will result in `/myrelative/myrelative/work_packages` if the base is `/myrelative`.

2. The process of parsing the URL is buggy and may not lead to the correct state.

Instead, I suggest to be explicit in what state this element should
route to.

https://community.openproject.com/projects/openproject/work_packages/25345