[25411] Never render extra parents in timeline
Currently, hierarchy mode does NOT load extra parents that are not part
of the results table (marked grey).

As they are not loaded, the timeline doesn't have any information to
draw them with.

However, IF the workPackages state _had_ the value of the work package
(e.g., because it's open in split view or it is part of a relation), it would still be
rendered.

This PR avoids rendering these rows at all times.