[25415] Add ancestor groups always when inserting extra rows
The hierarchy render pass uses ancestor groups to identify which rows
belong to a certain hierarchy group for simple selector-based
hiding/expanding of nested groups.

For extra added rows, these groups were only added when the ancestor was
not yet added to the list, causing them to not respond to collapsing the
ancestor group.