have sums listen for changes
Before, the sums where initialized once and did not listen for updates. They used to listen for updates before the display directive was rebuild to make use of the wp cache service. As the sums are not placed in the cache, although they could have their own state in the long run, we need to watch ourselves for changes in case of a custom schema.