[25432] Correctly transmit hasDefault to frontend
The default value for CFs is '', which is obviously not nil despite not
being a default, however the frontend only _always_ shows the attribute
when it is required and has no default.

This PR also adds a 'always visible' flag to the form configuration when
the CF is set required.