Implement query zoom handling in frontend
Had to refactor:

1. The zoom level enum in the frontend cannot properly be matched to
strings so I chose a separate type instead to map to the string
retrieved from /sent to the query

2. The way query updates are done through the states. The previous
implementation only worked for one single attribute state.

For the timeline state, it does not make sense to split that into two,
so I delegate to the state to handle checking and updating the query.