[25438] Fix only saving changed column positions
When changing only the positions of the default query columns,
the query incorrectly checked only for the permutation of the default
columns, not the actual order.

Thus the columns were not actually saved in the query.

Additionally, the `observeUntilChanged` incorrectly reflects previous
values which causes the movement of a column twice to only result in a
move by one position.

As this was already suggested to remove in
https://github.com/opf/openproject/pull/5597, let's resolve this
together in this PR.