avoid side effects on calls to attribute_groups (#5677)
Without the deep_dup, the call to attribute_groups used to alter the type's attribute_groups deeply nested attribute_groups property. As a result, the second call to attribute_groups would be different from the first one. This became most apparent when rending the schema for a couple of work packages all sharing the same type. As the type was reused to render the schema, but the project might have changed, only the first schema for the type was accurate, while subsequent schemas for that type lacked attributes

[ci skip]