[25615] Render content of additional hierarchy rows
Renders additional hierarchy elements as regular citizens of the table.
To this end, we need to load all ancestors when the hierarchy mode is
enabled.

This PR combines the WP loading of the relation columns service with the
required WPs of additional hierarchies into a separate service.

Additionally, whenever the hierarchy mode is turned on, we now need to
visibly refresh the results.