fix author filter for projectless users (#5718)
As the principals returned by the principals API where limited to those that are members in projects the user can see, users not having any project would not be returned. This is intended in so far as to prevent getting all users easily. But the user herself should be visible in any case.

[ci skip]