[25707] Avoid merging all involved state values, but only those requested
We cannot ensure that the relations of involved work packages we requested relations
for are complete. We only know that for the set of _requested work
packages_, all visible outgoing relations are shown.

We can thus only fill those work package states, and keep the others
empty. If an involved work package requires their relations (e.g.,
timeline of expanded relation column or split view), they are requested
on demand.