reload newly created wp after file upload completed (#5790)
The reload is necessary to ensure that attachments, referenced to be embedded inside the description can be correclty referenced in the backend when the textile rendering happens. That way, the embedded images are correctly displayed after the file upload completes.

There might be a phase where broken images are displayed. I find this acceptable and refrained from improving the solution.

The solution itself is a hack working around the wrong ordering of the file upload and the work package creation. As such, I didn't deem it necessary to improve the solution any further. E.g. the reload happens regardless of whether it is actually necessary as the code does not check whether an image is actually referenced. The only check is for the existence of an attachment

[ci skip]