Store only the changeset in a state, not the form itself
With the changeset being the authoritive source of the changes to a work
package, we can pass it around.

Also, a temporary resource for viewing the changes made to the work
packages is available through the editing service.

This is required since we need to show the WP resource as it were if the
changes were applied.