[26035] Don't clear stale cache states
When a state is stale, its value was cleared.
This causes e.g., a visual delay in the split screen beacuse a work package is no longer
present in the system.

If we instead only load the value, the stale value will be available for
the period of the request of the next one.

Regular value-subscriber of the state will thus receive the stale value
first, followed by the refreshed value.

https://community.openproject.com/wp/26035