[26044] Handle escape in field handler and reference form in table
Mousetrap was used to trigger escape on the field, but since we already
have a keydown handler attached to it, we can re-use it for cancelling
the field editing.

While fixing this, I found out that the table edit context uses a new
form for each field, which means the edit handler of the previous field
is lost. A new reference context is held in the table to return existing
forms, if any.
In the single view, this is tied to the $scope of the edit field group,
and thus ensure only one form exists.