Replace scopes with already defined names
The activerecord scope helpers output a warning when a scope's
name already exists in the module's method chain.

E.g., `Kernel#open` is overridden by some of the scopes, but that always
results in a warning.

A better way is to explicitly override the method in the model using a
class method.