add link to filter by lft
on lft sorting, projects are ordered first hierarchically, then alphabetically. However, this is faked as the sorting is only applied to the projects returned by the sql ("ORDER lft asc"). So if 100 projects are returned, only those 100 are sorted alphabetically while the 101 element could potentially belong inside that result set.

As lft and the other criteria are mutually exclusive, the UI will remove lft when sorting by one of the other criteria and all other criteria when sorting by lft