hack availability of wp id filter
The id filter is currently implemented as a stub. It is usable from the wp api but using it via the query api and by that via the front end fails intentionally as we currenlty have no solution to display the potentiall filter values efficiently. As there could be thousands of possible values, this problem is quite hard.

This hack circumvents the problem by not showing the filter in the query representer.

The filter is applied but the user will not be informed of the existence of the filter or its values.

As this is more of an easter egg, this is probably ok for now.