[26772] Fix double display of hierarchy elements
Assume the following work package structure

|ID      | Subject     | Type |
|--------|-------------|------|
|N       | Child       | Task |
|N+1     | Parent      | Bug  |
|N+2     | Grandparent | Task |

And the following query:

```
Filter Type = Task
Sort by ID asc
Hierarchy mode ON
```

You would see the ID N+2 Grantparent twice in the table

```
V Grandparent
.. V Parent
.... Child
V Grandparent
```

This was due to the grandparent being in the filter results, and the
deferral check of the hierarchy
mode applying only to the immediate parent. That parent is not contained
in the list, so the
elements were rendered immediately.

https://community.openproject.com/wp/26772