allow AR to alias on distinct
When passing a string to #order AR has no chance to alias columns. This can clash when two columns from different tables having the same name are part of the query (e.g. id). By providing an array, AR applies aliases to columns it needs to have in the projection on SELECT DISTINCT ... ORDER BY statements.