Avoid require memory leak with pathnames in load path (#6148)
Any ruby requires that occur in your application will leak with
Pathnames in the $LOAD_PATH in ruby 2.3/2.4. Beyond that, it's faster
to use strings in the load path. Please join this bug and add your
support for a fix:

https://bugs.ruby-lang.org/issues/14372

Same problem found here: https://github.com/lobsters/lobsters/pull/449

[ci skip]