return proper WWW Authenticate header on missing credentials
An anonymous_fallback is introduced and registered to be the last
strategy for warden. That strategy will always apply and it will always
return the anonymous user.

The better way would be to only apply the strategy if login is not
required. That way we would still be able to return a 403 when
credentials are missing but would no longer need to have 401 handlers
outside of warden. That duplicity caused the missing header in the first
place.

It would however require us to handle the realm in the strategy and by
that increase the coupling between the strategy and the application.