Avoid DatabaseCleaner truncation strategy with in-process puma server
Since Rails 5.1., we can force the application server and capybara to
share the same connection, which results in database_cleaner no longer
being necessary and resulting in faster tests due to transactions.

This will only work with a server that works in single mode (e.g., Puma).
We're still using database_cleaner in cucumber however, so we can use it to ensure we run with a clean database (which use_transactional_fixtures does not ensure) by starting it with truncation once.

This PR also adds necessary wiring to support `system` tests, however I don't see the immediate benefit there. rspec will continue to support the previous feature specs syntax.