Allow :none authentication as empty smtp_authentication setting (#6210)
We have users that need SMTP without authentication present, e.g., in an internal network. When passing :none to the smtp settings, this results in an authentication error since it tries to use that authentication method.

The correct way is to remove the authentication setting in order to use none. https://stackoverflow.com/a/26093613

https://community.openproject.com/wp/27284

[ci skip]