use integer for primary key columns
Rails 5.1 uses bigint(20) as the default for primary key columns (id typically) when it used to be int(11) up until Rails 5.0. While it might make sense to change to bigint in the long run, we can not do this by altering the old migrations as existing installations would not profit from the change.

We therefore force using int(11) for all old, and potentially already executed, migrations.