improve performance of relation query
By restructuring the `or` to be within the subselect, the query engine is able to filter down on the relations way more efficient than before.

As this query is used within WP scheduling, changes to the dates are more performant now. On my setup with about 500k relations on MySql, the change saved about 500ms