introduce custom eager loading for work package list
Reduces the amount of AR models instantiated when handing wp#index
requests by:

* Calculating part of the checksum for caching in the db. This is done
by using the id and the updated_at/on field of associated models whose
representation is cached in the wp representer.
* Reusing instantiated models for multiple associations. e.g instances
of Project are used in various places (the work package itself, it's
ancestors and children). Instead of using Rails' eager loading which
would instantiate a project potentially once for every association path,
the code now assigns the same instance to all those associations.

Additionally, the commit streamlines the creation of custom field
injected representers. Now only one class is created for all work
packages when before it would be one class for every uniq set of custom
fields. On top of that, the class can be reused if the set of custom
fields does not change.