Rework helper to auto-include wysiwig in tab builder when requested
This fixes the position of the wysiwig editor when used in combination
with the tab builder.

This requires being more clever with the helpers rather than simply
including them everywhere.

To this end, we require the helpers to be passed the current view
context (or template) to render the necessary bits.