[27600] Add link to query_form to represent allowing new queries to be saved
When an existing query (and its form) is loaded, members with
permissions to create new queries, but NOT to update existing ones can
not save existing queries as new ones.

This is due to the `create`, `update` and `commit` (on form) links are
dependent on the query resource itself, not on general permissions of
the queries project.

https://community.openproject.com/wp/27600