[27462] Pass wiki title to new page if it does not exist
When creating a wiki page from the text with, e.g., `[[Foo Bar]]`,
the user gets redirected to wiki#show for a page `foo-bar` (the slug of
the title above).

He thus loses the title attribute he entered previously. Instead we can
pass the title to the new wiki page IFF the page does not exist.

https://community.openproject.com/wp/27462