[28041] Allow CKEditor preview functionality to expand macro components
Even while we're using AOT, we can dynamically expand known components
that are meant for embedding (such as macros) to be dynamically
bootstrapped even on dynamic content such as previews.

But, we need a listing of what these components and their selectors are.
Very similar to how dynamic bootstraping works.

To this end, I have refactored the dynamic bootstrapping into a static
lookup class. This class can also output all components that are allowed
for embedding into dynamic contents.

The preview plugin is extended to use the new preview service that
bootstraps components within the view.

https://community.openproject.com/wp/28041