[28099] Close active edited resource when changing type
When the table has a required CF for the current type A,
and the required CF is open,
and you switch to a type B where this required CF does not exist,
the field does not close, but remains open, trying to access its field
schema. This field schema does no longer exist (no schema entry for this
CF).

https://community.openproject.com/wp/28099