[28167] Allow embedded table macros in WP description
This PR renders macros within WP descriptions. This has one drastic side
effect when embedded tables are rendered:

1. If the work package (lets asumme that ID #1) is returned in the
embedded table results, it is updated in the state/cache service.

2. The updated entry in the cache service results in refreshing the display fields

3. The display field will re-render the embedded table

4. Results are being re-fetched. This results in recursion into step 1

We can avoid that by assumming identical sources returned from the
backend to be identical work packages and in turn not updating those.
It's a hack, however.

https://community.openproject.com/wp/28167