allow adding attachments to invalid containers
Adding attachments does not change the validity of the container but we want to save the container to generate a journal when an attachment has been added. We therefore do not expect the user to fix the error on the container before them being able to attach to it again.