[28448] Avoid running into nil project on validation check
The `Subproject` filter controls through `applicable?` that it is not
applied to a valid subset when the given project is nil.

However, when testing an existing query where the project is no longer
existing (deleted, e.g.,), the filter validation will try to access
`values`, which in this case runs into a nil project.

We may want to consider moving this into a higher place, e.g., not
running all validations when the filter is not `applicable?`

https://community.openproject.com/wp/28448