Consolidate edit fields with their components
Previously, we have edit field modules (typescript classes) and their component counterparts (projected into CDK portals). This was mainly due to the old AngularJS solution used one edit field component with a dynamic field for rendering, which is now no longer applicable. In Angular, we need a 1-to-1 mapping between component and template for edit fields.

This PR thus treats `EditFieldComponents` as a subclass of `Field` so we don't need to build both. This issue became apparent in the CKEditor instance.