avoid adding non writeable filters multiple times
When the create_class method is called on WorkPackagePayloadRepresenter, a subclass of the class is created that includes the relevant custom fields. That class gets the PayloadRepresenter module included. The payload representer adds a render_filter to the attributes to avoid rendering unwriteable attributes. But when adding the filter, it is not added to an instance of the attribute in the subclass (WorkPackagePayloadRepresenter + CustomFields)but rather to the one in the superclass (WorkPackagePayloadRepresenter). The filter list instance is shared by all subclasses. That sideeffect is not dangerous per se as we want all WorkPackagePayloadRepresenters to have the filter. But as we have one instance of an array, adding to that array will lead to having the filter multiple times which will then all have to be evaluated on every rendering pass. By checking if the filter is already included before adding it, we can avoid this behaviour.

The better fix would be to rework the way custom fields are handled in the representers. This could be done by having one representer for the static attributes and one for the dynamic custom fields. Those two representers could be rendered/parsed independently and the results could then be merged. This approach however takes considerably more time.