Consolidate edit fields with their components (#6688)
* Consolidate edit fields with their components

Previously, we have edit field modules (typescript classes) and their component counterparts (projected into CDK portals). This was mainly due to the old AngularJS solution used one edit field component with a dynamic field for rendering, which is now no longer applicable. In Angular, we need a 1-to-1 mapping between component and template for edit fields.

This PR thus treats `EditFieldComponents` as a subclass of `Field` so we don't need to build both. This issue became apparent in the CKEditor instance.

* Better error handling to avoid JSON errors on submit

* Restore onSubmit we need to use to update field values

* Fix new spec by correctly rejecting save

* Fix demo data seeder spec

* Correctly read data from onSubmit handler before saving

Otherwise, the value saved may not be the absolute recent value.