[29125] Allow multiple sessions with remember tokens
Whenever creating a new remember token, all other previous tokens where automatically invalidated.
This results in cumbersome logins for users with multiple browsers.

Instead, this PR allows multiple tokens, but shows to the user a notice whenever ANY remember token exists for his account,
even when it is not in the current session.

This will allow the user to invalidate *all* remember tokens.

https://community.openproject.com/wp/29125