Fix reloading wp-table on global query resource
We had a split query and table state for a while during the split global
and local table state for migration reasons.

We no longer use that so we can safely put the query resource and form
into the table state, since that should be localized for each embedded
table.

This avoids table reload errors caused by an empty query, which is now
only loaded in the local state.