Ensure sql join in manual filter and sort column is isolated in query
We depend on ordered_work_packages association
for determining sort and filter for manual sorting.

We could restrict the join result with where(query_id: context.id) later
but that prevents the execution planner from optimizing on the explicit join clause.

This PR introduces a sql join that depends on the query_id and thus
retrieves the correct order performantly.