handle permissions on systems with fs.protected_regular=1
On newer Linux systems (>4.19), the kernel parameter
fs.protected_regular=1 prohibits writing to files which do not belong to
the user in world writable sticky directories like /tmp.

Due to the way openproject writes $PGPASSWORD to a temp file in /tmp,
the dockerised openproject fails with:

  /usr/src/app/docker/entrypoint-all-in-one.sh: line 20: /tmp/tmp.SYq0Do3sUb: Permission denied

This commit first writes the password as root to the file, then changes
the ownership to postgres, and finally deletes the file as postgres.

[ci skip]