application loading for rake tasks when no db exists
Since rails 5, the whole of the application is loaded (including all initializers and models) when a rake task is to be executed. In case no database exists, e.g. when issuing a db:create, this runs into errors in case calls to ActiveRecord methods require the existence of a database. Using the null-db gem, we can mimick the existence of a db in such cases.

Rails seems to reset the database connection to the one defined in the config/database.yml by itself.