[30137] Don't close the context menu on bubbling right click
Firefox sends contextmenu and click events independently and thus the
right click event may not be halted correctly.

To ensure the menu is not closed too fast, we can either wait for a
longer time before unsetting `isOpening`, or what this PR suggests,
avoid closing the menu on right clicks.

https://community.openproject.com/wp/30137

[ci skip]