use default filter behaviour for wps available for relations
That way, we can treat the end point as just another wp endpoint with
filter capabilities. Currently, there is a couple of custom logic still
present but that is only due to wanting to not break the public API.

In the long run, the whole end point could be removed as the same
filtering is now available using /api/v3/work_packages with an
appropriate filter.