prevent useless edits to wp status
In case the fixed version and the status of a work package are both closed, the status of a work package cannot be altered any more. Attempting it will result in an error.

Before the fix, the possible status calculation would not factor in that situation and the status would be marked as writable regardless as well. Therefore, the frontend would display a list of statuses even though selecting one would do nothing at all.

Now, the status property is marked as non writable in such a situation and the status list will only contain the current status. As we only fetch the form after a click on the edit field, we can only disable the field after that which is a bit weird.