[30763] Use stable sort for manual position
Only moved items are being sorted, to avoid sorting the entire page when
that is not yet necessary. This is a design decision to ensure only the
minimum number of positions are being assigned and helps when sorting a
restricted set of work packages from a bigger set.

To avoid having ambiguous sort results, ensure a second sort parameter
by ID.

https://community.openproject.com/wp/30763