[30777] Replace most usages of refresh service with global eventing service
We are increasingly having one problem: Notifying a parent querySpace of
changes within another embedded query space (e.g., a nested embedded
children table adding new relations, which should re-render the table).

As the `WorkPackageTableRefreshService` is rightfully part of the query space,
it cannot be used for these notifications.

This PR suggests to add a global `WorkPackageEventService` which is
called for specific events that can be registered to and events can be
received in buffered mode with a debounce. From these events, the table
is then refreshed if a visible work package is present in the table.

https://community.openproject.com/wp/30777