Don't break session auth if fallback session provided
When no session is provided (e.g., just curl an attachment URL),
Rack will still give you a session that is empty, which of course has
TTL attribute outdated if being checked for it.

Thus we accept an empty placeholder session as valid as long there are
no attributes. The only user authenticatable through it is
`User.anonymous`.