basic height calculation for graphs
Have a relatively positioned container for graphs as prescribed (https://www.chartjs.org/docs/latest/general/responsive.html#important-note) to allow basic resizing to automatically work. For horizontal bars, we calculate the height as well as it otherwise tends to become rather small, to the point of it becoming unreadable.