add type attribute group attributes to cache key
This is an attempt to fix cases where some custom fields are not displayed. I suspect this will not fix the bug in itself. But it might help to narrow down the problem by preventing the bug to also affect users who are not affected by the bug originally but only because the buggy result was cached for them.