actually provide all relevant custom fields for performance hack
Before, the last iteration on a project determined the custom fields that where hacked to be loaded in a project. That way, when the last project had e.g. no custom fields on a specific type, the custom fields would be missing for all evaluations later on, even though it might actually have custom fields defined for other types. As the results of the first call would be cached, this would then also affect later requests