[30994] Update empty changesets to match the latest resource
If a work package changeset is openend, and then some other action
happens that does not run through a changeset (updating a parent id
e.g.,), the old changeset will still be present and returned by the
editing service. This results in getting an older lockVersion for the
edit, even though the work package is up to date.

https://community.openproject.com/wp/30994