[31082] Remove max-height from ckeditor
As we now have a proper `position: sticky` solution, the max-height is
no longer required. It can by itself scroll when the next outer scroll container
is being scrolled, which works on both content-wrapper scrolled pages,
and work package pages.

However that requires that we avoid `overflow: hidden` on the
`form--fields` so this introduces a modifier for globally scrolled pages
to do just that.

https://community.openproject.com/wp/31082