[31085] Add missing gon settings to my page
All pages that render the `openproject-base` tag in a layout must pass
the gon settings.

This is actually a bad pattern that could be solvable with the
configuration api endpoint. It is being loaded for every work package
page load, but does not contain the relevant information such as comment
sorting.

In fact, it is only being used for the pagination component and differs
in structure from the gon settings.

So this should be considered a quick fix for 10.0. with a proper
solution (removing gon settings) to follow in
https://community.openproject.com/wp/31089

https://community.openproject.com/wp/31085