[31094] Remove hidden overflow when dropdown opens
It causes the table to shift due to the removal of the scroll bar which
then triggers a reflow. In many cases this takes longer than the
reposition if many work packages are shown.

As a workaround for 10.0., we DO allow scrolling the outer container but
close the dropdown as soon as it is being scrolled.

https://community.openproject.com/wp/31094