Add special case for new descriptions again
The new changeset is meant to avoid ever writing to work packages. In
order to take over default descriptions on type change however, we need
to avoid:

- overriding a description that already exists in the changeset
- writing the description in the changeset because that will
result in the description sticking around when switching types again
without ever typing in the description

The previous solution worked because an actual work package resource was
generated from scratch when type or project was changed. We no longer do
this with the proxy. There's no longer any resource where we can change
the description other than the original work package.

For new work packages, this doesn't matter as this is destroyed and
recreated anyway. Thus for new work packages, we can solve the
description setting from the form's default.