[31123] Only cast numeric custom values
It turns out that community still has some custom values that have not
been turned into custom_options with their values a numeric reference.

This results in trying to cast string values that cannot be casted as
numeric. We can avoid this without migrating all custom values again by
checking their format with a regex. This is slower as executed on every
sort, but will only affect multi-select list custom field sorting, which
is one sort criteria on the query:

https://community.openproject.com/projects/openproject/work_packages?query_id=1910

https://community.openproject.com/wp/31123