Allow `deliver_later` using a custom ApplicationJob
We're getting more and more reports on synchronous Timeouts and SMTP
errors causing internal errors for users when trying to send mail in the
request of the browser.

With rails 5.2, we can assign a custom job that handles the
`deliver_later` delayed sending. We can hook this into the
ApplicationJob with delayed_job.

Since rails now also has GlobalID serialization of ActiveRecord, we
don't even need to do the serialization ourselves!