rely on api for configuration instead of gon
Replaces all configuration fields currently supported by the api to be fetched from the api instead of from gon. Reduces the information passed to gon. Only the date/time format fields are left for gon to pass as those are not yet available via the api. Adding the fields should be easy.