Remove from documentation misleading traces of MySQL
This commit can be considered Hotfix, as it covers documentation bugs.
Since version stable/10, MySQL is not available anymore for installation.
However, there are plenty of traces of MySQL on the documentation itself.

The associated PR is still Requested for Comments and review, there is one
unclear point: pertinence and format of the `DATABASE_URL` env variable.

Maybe iterating to better explain globally the `openproject configure` se-
-ction, all env var indexes with explanation and location of conf files.

Tested and "compiled" successfully on my forked repo (Github display MD).