fix dragging to outside of grid
AS the code relied on the mousedOverArea, and that area was not set when moving outside of the grid or over another widget, the drop failed. As we no longer rely on the drop even being fired by the receiving grid-area but listen to the drop event on the moved widget area we can get rid of the stop() function which used to complicate things. By that, we can also simplify the abort case and no longer need to maintin the magic `aborted` variable