[31504] Reload attachments again after coming from list
Due to the way the attachments are embedded in HAL, we can end up with
two situations:

1. We load the work package in single-view and attachments are embedded
($loaded = true)

2. We load the work package in the list and attachments are not embedded
($loaded = false)

The attachment-list doesn't properly reload the attachments in the first
case after being initialized. Assume you open the single-view, upload an
attachment and then let the list reload, the work package attachments
will become unloaded since they are contained in the work package
resource and not in a separate state.

We thus have to reload attachments more often to ensure they get
reloaded in this case. The proper solution would be to have attachments
tracked in a separate resource.

https://community.openproject.com/wp/31504