[31583] avoid parsing strings to yaml
If a configuration/ENV value has YML-incompatible contents, it is broken
due to the `YAML.load` approach.

However, if the outcome of `YAML.load` is a string, we can expect the
input to be an actual string, not something YAML should parse.

Thus as a workaround until proper configuration types are added (or
better, settings and configuration is joined into one), this avoids
breaking passwords in ENV configuration inputs.

https://community.openproject.com/wp/31583