prevent involuntary modal close
When clicking to remove error messages or on the labels e.g. in the registration modal, the elements clicked on are for unkonwn reasons not part of the DOM. Thus, the check for whether the click was performed inside the modal fails even thoug the click was performed inside the modal at least from a visual stand point.

Changing the default fixes the problem and does not seem to cause problems with the other modals.