do not fail pdf export on unpermitted image url
In case the description contains an image, prawn_markup attempts to download the image to embed it in the html turned pdf. But in case of errors, the gem does not seem to them thouroughly. E.g. a 403 when fetching the image, causes the whole export to fail. It might be worthwile to further invest in the area to display certain images, but for the time being, we simply do not fetch the images any longer.