Use text pattern for float number fields
The number pattern doesnt correctly respect the lang field in all
browsers it appears. Using text field with
custom parsing/formatters with validations for the numbers, we can avoid the number field.

We however need to listen for keydown events to emulate a number field
and disable inputting any other fields. This will probably break for
some keys.